feat : add CRUD for board

+++ backend/src/main/java/com/cloudmall/backend/board/BoardRepository.java
@@ -2,6 +2,6 @@ package com.cloudmall.backend.board;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
+public interface BoardRepository extends JpaRepository<BoardEntity, Integer> {
 
 }
